refactor: split resolve-scoped-slot into its own file
