+++ src/core/instance/render-helpers/index.js
@@ -9,7 +9,7 @@ import { checkKeyCodes } from './check-keycodes'
 import { bindObjectProps } from './bind-object-props'
 import { renderStatic, markOnce } from './render-static'
 import { bindObjectListeners } from './bind-object-listeners'
-import { resolveScopedSlots } from './resolve-slots'
+import { resolveScopedSlots } from './resolve-scoped-slots'
 import { bindDynamicKeys, prependModifier } from './bind-dynamic-keys'
 
 export function installRenderHelpers (target: any) {
